% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebv_properties.R
\docType{class}
\name{EBV netCDF properties-class}
\alias{EBV netCDF properties-class}
\title{EBV netCDF properties class (S4)}
\value{
S4 class containing the EBV netCDF properties
}
\description{
EBV netCDF properties class (S4)
}
\section{Slots}{

\describe{
\item{\code{general}}{Named list. Elements: title, description, ebv_class, ebv_name,
ebv_subgroups, creator}

\item{\code{spatial}}{Named list. Elements: wkt2, epsg, resolution, extent, dimensions}

\item{\code{temporal}}{Named list. Elements: units, t_delta, timesteps,
timesteps_natural}

\item{\code{metric}}{Named list. Elements: standard_name, description}

\item{\code{scenario}}{Named list. Elements: standard_name, description}

\item{\code{ebv_cube}}{Named list. Elements: standard_name, description, unit, type,
fillvalue}
}}

\note{
If the properties class holds e.g. no scenario information this is
indicated with an element called status in the list.
}
