% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebv_attribute.R
\name{ebv_attribute}
\alias{ebv_attribute}
\title{Write a new attribute value to an EBV netCDF}
\usage{
ebv_attribute(
  filepath,
  attribute_name,
  value,
  levelpath = NULL,
  verbose = FALSE
)
}
\arguments{
\item{filepath}{Character. Path to the netCDF file.}

\item{attribute_name}{Character. Name of the attribute that should be
changed.}

\item{value}{New value that should be assigned to the attribute.}

\item{levelpath}{Character. Default: NULL. Indicates the location of the
attribute. The default means that the attribute is located at a global
level. If the attribute is located at the datacubelevel just add the
datacubepath, e.g. metric_1/ebv_cube. For the metric level the value may be 'metric_1' or
'scenario_1/metric_1'. This path depends on whether the netCDF hierarchy
has scenarios or not.}

\item{verbose}{Logical. Default: FALSE. Turn on all warnings by setting it to
TRUE.}
}
\value{
Adds the new value to the attribute. Check your results using
\code{\link[=ebv_properties]{ebv_properties()}}.
}
\description{
Write a new attribute value to an EBV netCDF. Not all attributes
can be changed. Some are always created automatically, e.g. the attributes
belonging to the crs, time and var_entity datasets. In this case you have
to re-create the netCDF file.
}
\note{
You can change the ebv_class and the ebv_name. In this case you need to
change the ebv_class first. Don't forget to change the ebv_name accordingly!
}
\examples{
file <- system.file(file.path("extdata","cSAR_new.nc"), package="ebvnetcdf")
attribute1 <- 'standard_name'
value1 <- 'mammals'
level1 <- 'metric01'
#ebv_attribute(file, attribute1, value1, level1)
attribute2 <- '_FillValue'
value2 <- -999
level2 <- 'metric01/entity01'
#ebv_attribute(file, attribute2, value2, level2)
attribute3 <- 'creator'
value3 <- 'Jane Doe'
#ebv_attribute(file, attribute3, value3)
}
