% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebv_data_read_bb.R
\name{ebv_data_read_bb}
\alias{ebv_data_read_bb}
\title{Read subset (bounding box) of the data from EBV NetCDF}
\usage{
ebv_data_read_bb(
  filepath,
  datacubepath,
  bb,
  outputpath = NULL,
  timestep = 1,
  epsg = 4326,
  overwrite = FALSE,
  ignore.RAM = FALSE
)
}
\arguments{
\item{filepath}{Path to the NetCDF file.}

\item{datacubepath}{Path to the datacube (use
\code{\link[=ebv_datacubepaths]{ebv_datacubepaths()}}).}

\item{bb}{Definition of subsset by bounding box: c(xmin, xmax, ymin, ymax).}

\item{outputpath}{Defaul: NULL, returns the data as a raster object in
memory. Optional: set path to write subset as GeoTiff on disk, returns
outputpath.}

\item{timestep}{Choose one or several timesteps (vector).}

\item{epsg}{Default: 4326 (WGS84). Change accordingly if your bounding box
coordinates are based on a different coordinate reference system.}

\item{overwrite}{Default: FALSE. Set to TRUE to overwrite the outputfile
defined by 'outputpath'.}

\item{ignore.RAM}{Checks if there is enough space in your memory to read the
data. Can be switched off (set to TRUE).}
}
\value{
Returns a raster object if no outputpath is given. Otherwise the
subset is written onto the disk and the ouputpath is returned.
}
\description{
Read a subset of one or more layers from one datacube of the
NetCDF file. Subset definition by a bounding box.
}
\examples{
# file <- 'path/to/netcdf/file.nc'
# datacubes <- ebv_datacubepaths(file)
# out <- 'path/to/write/subset.tif'
# bb_utm32 <- c(271985, 941837, 5232640, 6101151)
# cSAR.germany <- ebv_data_read_bb(file, datacubes[1], c(5,15,47,55), timestep = c(1,4,12))
# path <- ebv_data_read_bb(file, datacubes[1], c(5,15,47,55), out, timestep = c(2,3))
# path  <- ebv_data_read_bb(file, datacubes[1], bb_utm32, out, timestep=1, epsg=32632, overwrite=T)
}
