% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebv_datacubepaths.R
\name{ebv_datacubepaths}
\alias{ebv_datacubepaths}
\title{Get datacubepaths}
\usage{
ebv_datacubepaths(filepath)
}
\arguments{
\item{filepath}{Path to the NetCDF file.}
}
\value{
Dataframe containing the paths to access the datacubes and
descriptions of scenario, metric and entity if existing.
}
\description{
Get the paths to the datacubes to access and read the data.
}
\examples{
# file <- 'path/to/netcdf/file.nc'
# datacubes <- ebv_datacubepaths(file)
}
