% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebv_indicator.R
\name{ebv_indicator}
\alias{ebv_indicator}
\title{Plot the average over time of one datacube of an EBV NetCDF}
\usage{
ebv_indicator(
  filepath,
  datacubepath,
  entity = NULL,
  color = "dodgerblue4",
  verbose = FALSE
)
}
\arguments{
\item{filepath}{Character. Path to the NetCDF file.}

\item{datacubepath}{Character. Path to the datacube (use
\code{\link[=ebv_datacubepaths]{ebv_datacubepaths()}}).}

\item{entity}{Character or Integer. Default is NULL. If the structure is 3D,
the entity argument is set to NULL. Else, a character string or single
integer value must indicate the entity of the 4D structure of the EBV
netCDFs.}

\item{color}{Character. Default: dodgerblue4. Change to any color known by R
\code{\link[grDevices:colors]{grDevices::colors()}}}

\item{verbose}{Logical. Default: FALSE. Turn on all warnings by setting it to
TRUE.}
}
\value{
Plots a line plot and returns a vector of the average. If the data
encompasses only one timestep a single mean is returned.
}
\description{
Plot the average (y-axis) of one datacube of a EBV NetCDF over
time (x-axis). If the datacube has only one timestep a single mean value is
returned.
}
\examples{
#set path to EBV netCDF
file <- system.file(file.path("extdata","cSAR_idiv_v1.nc"), package="ebvcube")
#get all datacubepaths of EBV netCDF
datacubes <- ebv_datacubepaths(file)

#plot the change of the mean over time of the first datacube
ebv_indicator(filepath = file, datacubepath = datacubes[1,1], entity = NULL)
}
