% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebv_ncdf_add_data.R
\name{ebv_ncdf_add_data}
\alias{ebv_ncdf_add_data}
\title{Add data to self-created EBV NetCDF}
\usage{
ebv_ncdf_add_data(
  filepath_nc,
  filepath_tif,
  datacubepath,
  timestep = 1,
  band = 1,
  ignore.RAM = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{filepath_nc}{Path to the self-created NetCDF file.}

\item{filepath_tif}{Path to the GeoTiff file containing the data.}

\item{datacubepath}{Character. Optional. Path to the datacube (use
\code{\link[=ebv_datacubepaths]{ebv_datacubepaths()}}).}

\item{timestep}{Default: 1. Define to which timestep or timesteps the data
should be added. If several timesteps are given they have to be in a
continuous order. Meaning c(4,5,6) is right but c(2,5,6) is wrong.}

\item{band}{Default: 1. Define which band(s) to read from GeoTiff. Can be
several. Don't have to be in order as the timesteps definition requires.}

\item{ignore.RAM}{Checks if there is enough space in your memory to read the
data. Can be switched off (set to TRUE).}

\item{verbose}{Logical. Turn on all warnings by setting it to TRUE.}
}
\value{
Adds data to the EBV NetCDF. Check your results using
\code{\link[=ebv_datacubepaths]{ebv_datacubepaths()}}, \code{\link[=ebv_data_read]{ebv_data_read()}} and/or
\code{\link[=ebv_data_analyse]{ebv_data_analyse()}}.
}
\description{
Add data to the self-created EBV NetCDF from GeoTiffs.
}
\note{
If the data exceeds your memory the RAM check will throw an error. No
block-processing or other method implemented so far. Move to a machine with
more capacities for the moment.
}
\examples{
# file <- 'path/to/created/netcdf/file.nc'
# tif <- 'path/to/geotiff/containing/data.tif'
# scenario <- 1
# metric <- 1
# entity <- 1
# ts <- c(1:6)
# band <- c(1:6)
# ebv_ncdf_add_data(file, tif, metric, scenario, entity, ts, band)
}
