% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebv_ncdf_create.R
\name{ebv_ncdf_create}
\alias{ebv_ncdf_create}
\title{Create an EBV NetCDF}
\usage{
ebv_ncdf_create(
  jsonpath,
  outputpath,
  entities.no = 0,
  epsg = 4326,
  extent = c(-180, 180, -90, 90),
  overwrite = FALSE
)
}
\arguments{
\item{jsonpath}{Path to the json file downloaded from the \href{https://portal.geobon.org/api-docs}{Geobon Portal API}.}

\item{outputpath}{Set path where the NetCDF file should be created.}

\item{entities.no}{Default: 0. Indicates how many entities there are be per metric.}

\item{epsg}{Default: 4326 (WGS84). Defines the coordinate reference system via the corresponding epsg code.}

\item{extent}{Default: c(-180,180,-90,90). Defines the extent of the data: c(xmin, xmax, ymin, ymax).}

\item{overwrite}{Default: FALSE. Set to TRUE to overwrite the outputfile defined by 'outputpath'.}
}
\value{
Creates the NetCDF at the 'outputpath' location.
}
\description{
Create the core structure of the EBV NetCDF based on the json from the \href{https://portal.geobon.org/api-docs}{Geobon Portal API}. Data and attributes will be added afterwards. See \code{\link[=ebv_ncdf_add_data]{ebv_ncdf_add_data()}} and \code{\link[=ebv_ncdf_entity_attributes]{ebv_ncdf_entity_attributes()}} for the next steps.
}
\note{
To check out the results take a look at your NetCDF file with \href{https://www.giss.nasa.gov/tools/panoply/}{Panoply}.
}
\examples{
# json <- 'path/to/json/file.json'
# out <- 'path/to/create/new/netcdf/file.nc'
# ebv_ncdf_create(json, out, 5)
}
