% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebv_ncdf_entity_attributes.R
\name{ebv_ncdf_entity_attributes}
\alias{ebv_ncdf_entity_attributes}
\title{Add entity attributes to self-created EBV NetCDF}
\usage{
ebv_ncdf_entity_attributes(
  filepath,
  datacubepath,
  standard_name,
  description,
  fillvalue = NULL,
  verbose = FALSE
)
}
\arguments{
\item{filepath}{Path to the self-created NetCDF file.}

\item{datacubepath}{Path to the datacube (use
\code{\link[=ebv_datacubepaths]{ebv_datacubepaths()}}).}

\item{standard_name}{Value of the standard_name attribute (character).}

\item{description}{Value of the description attribute (character).}

\item{fillvalue}{Value of the fillvalue attribute (double).}

\item{verbose}{Logical. Turn on all warnings by setting it to TRUE.}
}
\value{
Adds attributes to NetCDF. Check results using
\code{\link[=ebv_properties]{ebv_properties()}}
}
\description{
Add standard_name, description and fillvalue to an entitiy of a
self-created EBV NetCDF. First use \code{\link[=ebv_ncdf_create]{ebv_ncdf_create()}} to create
a NetCDF afterwards use this functiong to add attributes.
}
\examples{
# file <- 'path/to/self/created/netcdf.nc'
# datacubes <- ebv_datacubepaths(file)
# ebv_ncdf_entity_attributes(file, datacubes[1,1], 'habitat', 'bog', fillvalue=-1)
}
