% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebv_plot_indicator.R
\name{ebv_plot_indicator}
\alias{ebv_plot_indicator}
\title{Plot the average over time of one datacube of an EBV NetCDF}
\usage{
ebv_plot_indicator(
  filepath,
  datacubepath,
  color = "dodgerblue4",
  verbose = FALSE
)
}
\arguments{
\item{filepath}{Character. Path to the NetCDF file.}

\item{datacubepath}{Character. Path to the datacube (use
\code{\link[=ebv_datacubepaths]{ebv_datacubepaths()}}).}

\item{color}{Character. Default: dodgerblue4. Change to any color known by R
\code{\link[grDevices:colors]{grDevices::colors()}}}

\item{verbose}{Logical. Default: FALSE. Turn on all warnings by setting it to
TRUE.}
}
\value{
Plots a line plot and returns a vector of the average. If the data
encompasses only one timestep a single mean is returned.
}
\description{
Plot the average (y-axis) of one datacube of a EBV NetCDF over
time (x-axis). If the datacube has only one timestep a single mean value is
returned.
}
\examples{
file <- system.file(file.path("extdata","cSAR_idiv_v1.nc"), package="ebvnetcdf")
datacubes <- ebv_datacubepaths(file)
ebv_plot_indicator(file, datacubes[1,1])
}
