% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebv_properties.R
\name{ebv_properties}
\alias{ebv_properties}
\title{Read properties of EBV NetCDF}
\usage{
ebv_properties(filepath, datacubepath = NULL, verbose = FALSE)
}
\arguments{
\item{filepath}{Path to the NetCDF file.}

\item{datacubepath}{Optional. Path to the datacube (use
\code{\link[=ebv_datacubepaths]{ebv_datacubepaths()}}).}

\item{verbose}{Set to True to obtain all warnings.}
}
\value{
S4 class containing information about file or file and datacube
depending on input.
}
\description{
Structured access to all attributes of the NetCDF file.
}
\examples{
# file <- 'path/to/netcdf/file.nc'
# datacubes <- ebv_datacubepaths(file)
# prop_file <- ebv_properties(file)
# prop_dc <- ebv_properties(file, datacubes[1,1])
}
