% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebv_taxonomy_app.R
\name{ebv_taxonomy_app}
\alias{ebv_taxonomy_app}
\title{EBV taxonomy Shiny App}
\usage{
ebv_taxonomy_app()
}
\description{
Load an EBV netCDF that holds taxonomic information to browse
the data and visualize it.
}
\note{
Works with all EBV netCDFs created with
\code{\link[=ebv_create_taxonomy]{ebv_create_taxonomy()}}.
}
