% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebvcube-package.R
\docType{package}
\name{ebvcube}
\alias{ebvcube}
\title{Working with netCDF for Essential Biodiversity Variables}
\description{
This package can be used to easily access the data of the EBV netCDFs which
can be downloaded from the \href{portal.geobon.org}{Geobon Portal}. It also
provides some basic visualization of the data.  Advanced users can build
their own netCDFs with the EBV standard using this package.
}
\details{
This package contains three main usecases: accessing the data, visualising
it and creating your own data in the EBV netCDF
standard.
}
