% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world_boundaries}
\alias{world_boundaries}
\title{Simple outlines of world countries}
\format{
A data.frame with 177 elements and geometry as WKT
}
\source{
Data downloaded from
\href{https://www.naturalearthdata.com/downloads/110m-cultural-vectors/110m-admin-0-countries/}{Natural
Earth}. Used version 4.0.0 and reduced attributes.
}
\usage{
world_boundaries
}
\description{
Simple outlines of world countries
}
\keyword{datasets}
