% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wrld_simpl}
\alias{wrld_simpl}
\title{Simple outlines of world countries}
\format{
A Spatial Polygons Data Frame with 246 elements
}
\source{
Data imported from the \href{https://cran.r-project.org/package=maptools}{maptools package}: data(wrld_simpl, package='maptools')
}
\usage{
wrld_simpl
}
\description{
Simple outlines of world countries
}
\keyword{datasets}
