% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebv_properties.R
\docType{class}
\name{EBV NetCDF properties-class}
\alias{EBV NetCDF properties-class}
\title{EBV NetCDF Properties class (S4)}
\value{
S4 class
}
\description{
EBV NetCDF Properties class (S4)
}
\section{Slots}{

\describe{
\item{\code{general}}{list. Elements: title, description, ebv_class, ebv_name, ebv_subgroups,
creator}

\item{\code{spatial}}{list.}

\item{\code{temporal}}{list.}

\item{\code{metric}}{list.}

\item{\code{scenario}}{list.}

\item{\code{entity}}{list. Elements: value_range}
}}

