% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebv_create.R
\name{ebv_create}
\alias{ebv_create}
\title{Create an EBV NetCDF}
\usage{
ebv_create(
  jsonpath,
  outputpath,
  entities_no = 0,
  epsg = 4326,
  extent = c(-180, 180, -90, 90),
  fillvalue = NULL,
  prec = "double",
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{jsonpath}{Character. Path to the json file downloaded from the
\href{https://portal.geobon.org/api-docs}{Geobon Portal API}.}

\item{outputpath}{Character. Set path where the NetCDF file should be
created.}

\item{entities_no}{Integer. Default: 0. Indicates how many entities there are
per metric.}

\item{epsg}{Integer. Default: 4326 (WGS84). Defines the coordinate reference
system via the corresponding epsg code.}

\item{extent}{Numeric. Default: c(-180,180,-90,90). Defines the extent of the
data: c(xmin, xmax, ymin, ymax).}

\item{fillvalue}{Numeric. Value of the missing data in the array. Not
mandatory but should be defined!}

\item{prec}{Character. Default: 'double'. Precision of the data set. Valid
options: 'short' 'integer' 'float' 'double' 'char' 'byte'.}

\item{overwrite}{Logical. Default: FALSE. Set to TRUE to overwrite the
outputfile defined by 'outputpath'.}

\item{verbose}{Logical. Default: FALSE. Turn on all warnings by setting it to
TRUE.}
}
\value{
Creates the NetCDF file at the 'outputpath' location.
}
\description{
Create the core structure of the EBV NetCDF based on the json
from the \href{https://portal.geobon.org/api-docs}{Geobon Portal API}. Data
and attributes will be added afterwards. Use
\code{\link[=ebv_add_data]{ebv_add_data()}} to add the missing attributes.
}
\note{
To check out the results take a look at your NetCDF file with
\href{https://www.giss.nasa.gov/tools/panoply/}{Panoply} provided by the
NASA.
}
\examples{
json <- system.file(file.path("extdata","1.json"), package="ebvnetcdf")
out <- file.path(system.file(package='ebvnetcdf'),"extdata","sCAR_new.nc")
#ebv_create(json, out, 3, fillvalue=-3.4E38)
}
