% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebv_data_analyse.R
\name{ebv_data_analyse}
\alias{ebv_data_analyse}
\title{Title}
\usage{
ebv_data_analyse(
  filepath,
  datacubepath,
  subset = NULL,
  timestep = 1,
  at = TRUE,
  epsg = 4326,
  numerical = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{na.rm}{}
}
\value{
pending
}
\description{
Title
}
\examples{
#pending
}
\seealso{
\code{\link[=ebv_data_read_bb]{ebv_data_read_bb()}} and \code{\link[=ebv_data_read_shp]{ebv_data_read_shp()}} for the usage of subsets
}
