% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebv_ncdf_entity_attributes.R
\name{ebv_ncdf_entity_attributes}
\alias{ebv_ncdf_entity_attributes}
\title{Add datacube attributes to an EBV NetCDF}
\usage{
ebv_ncdf_entity_attributes(
  filepath,
  datacubepath,
  standard_name,
  description,
  fillvalue = NULL,
  verbose = FALSE
)
}
\arguments{
\item{filepath}{Character. Path to the self-created NetCDF file.}

\item{datacubepath}{Character. Path to the datacube (use
\code{\link[=ebv_datacubepaths]{ebv_datacubepaths()}}).}

\item{standard_name}{Value of the standard_name attribute.}

\item{description}{Character. Value of the description attribute.}

\item{fillvalue}{Numeric. Value of the fillvalue attribute.}

\item{verbose}{Logical. Default: FALSE. Turn on all warnings by setting it to
TRUE.}
}
\value{
Adds attributes to NetCDF at datacube level. Check results using
\code{\link[=ebv_properties]{ebv_properties()}}
}
\description{
Add standard_name, description and fillvalue to an
entitiy/datacube of a self-created EBV NetCDF. First use
\code{\link[=ebv_ncdf_create]{ebv_ncdf_create()}} to create a NetCDF afterwards use this
function to add attributes.
}
\examples{
file <- system.file(file.path("extdata","cSAR_new.nc"), package="ebvnetcdf")
# datacubes <- ebv_datacubepaths(file)
sn <- 'non forest birds species'
desc <- 'Changes in bird diversity at the grid cell level caused by land-use'
fv <- -3.4E38
#ebv_ncdf_entity_attributes(file, datacubes[1,1], sn, desc, fv)
}
