% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebv_ncdf_entity_attributes.R
\name{ebv_ncdf_entity_attributes}
\alias{ebv_ncdf_entity_attributes}
\title{Add entity attributes to self-created EBV NetCDF}
\usage{
ebv_ncdf_entity_attributes(
  filepath,
  datacubepath,
  long_name,
  label,
  units,
  fillvalue = NULL
)
}
\arguments{
\item{filepath}{Path to the self-created NetCDF file.}

\item{datacubepath}{Path to the datacube (use \code{\link[=ebv_datacubepaths]{ebv_datacubepaths()}}).}

\item{long_name}{Value of the long_name attribute (character).}

\item{label}{Value of the label attribute (character).}

\item{units}{Value of the units attribute (character).}

\item{fillvalue}{Value of the fillvalue attribute (double).}
}
\value{
Adds attributes to NetCDF. Check results using \code{\link[=ebv_properties]{ebv_properties()}}
}
\description{
Add long_name, label, units and fillvalue to an entitiy of a self-created EBV NetCDF. First use \code{\link[=ebv_ncdf_create]{ebv_ncdf_create()}} to create a NetCDF afterwards use this functiong to add attributes.
}
\examples{
# file <- 'path/to/self/created/netcdf.nc'
# datacubes <- ebv_datacubepaths(file)
# ebv_ncdf_entity_attributes(file, datacubes[1,1], 'habitat', 'bog', 'Percentage', fillvalue=-1)
}
