% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebv_plot_map.R
\name{ebv_plot_map}
\alias{ebv_plot_map}
\title{Map plot of an EBV NetCDF}
\usage{
ebv_plot_map(
  filepath,
  datacubepath,
  timestep = 1,
  countries = TRUE,
  col_rev = TRUE,
  classes = 5,
  ignore_RAM = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{filepath}{Character. Path to the NetCDF file.}

\item{datacubepath}{Character. Path to the datacube (use
\code{\link[=ebv_datacubepaths]{ebv_datacubepaths()}}).}

\item{timestep}{Integer. Choose one timestep.}

\item{countries}{Logical. Default: TRUE. Simple country outlines will be
plotted on top of the raster data. Disable by setting this option to FALSE.}

\item{col_rev}{Logical. Default: TRUE. Set to FALSE if you want the color
ramp to be the other way around.}

\item{classes}{Integer. Default: 5. Define the amount of classes (quantiles)
for the symbology. Currently restricted to maximum 15 classes.}

\item{ignore_RAM}{Logical. Default: FALSE. Checks if there is enough space in
your memory to read the data. Can be switched off (set to TRUE).}

\item{verbose}{Logical. Default: FALSE. Turn on all warnings by setting it to
TRUE.}
}
\value{
Plots a map.
}
\description{
Map plot of the data of one timestep in one datacube of an EBV
NetCDF. This functions sometimes writes temporary files on your disk.
Speficy a directory for these setting via
options('temp_directory'='/path/to/temp/directory').
}
\note{
Uses the country outlines data from the
\href{https://cran.r-project.org/package=maptools}{maptools package}.
}
\examples{
file <- system.file(file.path("extdata","cSAR_idiv_v1.nc"), package="ebvnetcdf")
datacubes <- ebv_datacubepaths(file)
ebv_plot_map(file, datacubes[1,1], timestep=9, classes=7)
}
