% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebv_properties.R
\name{ebv_properties}
\alias{ebv_properties}
\title{Read properties of EBV NetCDF}
\usage{
ebv_properties(filepath, datacubepath = NULL, verbose = FALSE)
}
\arguments{
\item{filepath}{Path to the NetCDF file.}

\item{datacubepath}{Optional. Path to the datacube (use ebv_datacubepaths()).}

\item{verbose}{Set to True to obtain all warnings.}
}
\value{
S4 class containing information about file or file and datacube depending on input.
}
\description{
Structured access to all attributes of the NetCDF file.
}
\examples{
filepath <- paste0(path.package("ebvnetcdf"),"/extdata/cSAR_idiv_v1.nc")
datacubepaths <- ebv_datacubepaths(filepath)
ebv_properties(filepath)
#prop_dc <- ebv_properties(file, datacubepaths[1,1])
}
